add LegacyInstance interface; update runtime packages
