+++ lib/runtime/interface.go
@@ -23,14 +23,34 @@ import (
 	"github.com/ChainSafe/gossamer/lib/trie"
 )
 
-// Instance is the interface a runtime instance must implement
+// LegacyInstance is the interface a legacy (v0.6) runtime instance must implement
+type LegacyInstance interface {
+	InstanceAPI
+	LegacyRuntimeAPI
+}
+
+// Instance is the interface a v0.8 runtime instance must implement
 type Instance interface {
+	InstanceAPI
+	LegacyRuntimeAPI
+
+	// TODO: parameters and return values for these are undefined in the spec
+	CheckInherents()
+	RandomSeed()
+	OffchainWorker()
+	GenerateSessionKeys()
+}
+
+type InstanceAPI interface {
 	Stop()
 	NodeStorage() NodeStorage
 	NetworkService() BasicNetwork
 
 	Exec(function string, data []byte) ([]byte, error)
 	SetContext(s Storage) // used to set the TrieState before a runtime call
+}
+
+type LegacyRuntimeAPI interface {
 	Version() (*VersionAPI, error)
 	Metadata() ([]byte, error)
 	BabeConfiguration() (*types.BabeConfiguration, error)
