+++ lib/runtime/wasmer/exports_unchecked_extrinsic_test.go
@@ -12,7 +12,7 @@ import (
 )
 
 func TestApplyExtrinsic_Transfer_NoBalance_UncheckedExt(t *testing.T) {
-	rt := NewTestInstance(t, runtime.NODE_RUNTIME)
+	rt := NewTestInstance(t, runtime.LEGACY_NODE_RUNTIME).Legacy()
 	rtVerB, err := rt.exec(runtime.CoreVersion, []byte{})
 	require.Nil(t, err)
 
@@ -76,7 +76,7 @@ func TestApplyExtrinsic_Transfer_NoBalance_UncheckedExt(t *testing.T) {
 }
 
 func TestApplyExtrinsic_Transfer_WithBalance_UncheckedExtrinsic(t *testing.T) {
-	rt := NewTestInstance(t, runtime.NODE_RUNTIME)
+	rt := NewTestInstance(t, runtime.LEGACY_NODE_RUNTIME).Legacy()
 	rtVerB, err := rt.exec(runtime.CoreVersion, []byte{})
 	require.Nil(t, err)
