+++ lib/runtime/wasmer/legacy_imports_test.go
@@ -27,7 +27,7 @@ import (
 )
 
 func Test_ext_twox_256(t *testing.T) {
-	instance := NewTestInstance(t, runtime.TEST_RUNTIME)
+	instance := NewTestInstance(t, runtime.TEST_RUNTIME).Legacy()
 	mem := instance.vm.Memory.Data()
 
 	data := []byte("hello")
@@ -54,7 +54,7 @@ func Test_ext_twox_256(t *testing.T) {
 }
 
 func Test_ext_kill_child_storage(t *testing.T) {
-	instance := NewTestInstance(t, runtime.TEST_RUNTIME)
+	instance := NewTestInstance(t, runtime.TEST_RUNTIME).Legacy()
 	mem := instance.vm.Memory.Data()
 	// set child storage
 	storageKey := []byte("childstore1")
@@ -109,7 +109,7 @@ func Test_ext_kill_child_storage(t *testing.T) {
 }
 
 func Test_ext_clear_child_storage(t *testing.T) {
-	instance := NewTestInstance(t, runtime.TEST_RUNTIME)
+	instance := NewTestInstance(t, runtime.TEST_RUNTIME).Legacy()
 	mem := instance.vm.Memory.Data()
 	// set child storage
 	storageKey := []byte("childstore1")
@@ -164,7 +164,7 @@ func Test_ext_clear_child_storage(t *testing.T) {
 }
 
 func Test_ext_get_allocated_child_storage(t *testing.T) {
-	instance := NewTestInstance(t, runtime.TEST_RUNTIME)
+	instance := NewTestInstance(t, runtime.TEST_RUNTIME).Legacy()
 	mem := instance.vm.Memory.Data()
 
 	// set child storage
