+++ lib/runtime/wasmtime/exports.go
@@ -26,12 +26,12 @@ import (
 )
 
 // Metadata calls runtime function Metadata_metadata
-func (in *Instance) Metadata() ([]byte, error) {
+func (in *LegacyInstance) Metadata() ([]byte, error) {
 	return in.exec(runtime.Metadata, []byte{})
 }
 
 // Version calls runtime function Core_Version
-func (in *Instance) Version() (*runtime.VersionAPI, error) {
+func (in *LegacyInstance) Version() (*runtime.VersionAPI, error) {
 	res, err := in.exec(runtime.CoreVersion, []byte{})
 	if err != nil {
 		return nil, err
@@ -51,7 +51,7 @@ func (in *Instance) Version() (*runtime.VersionAPI, error) {
 }
 
 // BabeConfiguration gets the configuration data for BABE from the runtime
-func (in *Instance) BabeConfiguration() (*types.BabeConfiguration, error) {
+func (in *LegacyInstance) BabeConfiguration() (*types.BabeConfiguration, error) {
 	ret, err := in.exec(runtime.BabeAPIConfiguration, []byte{})
 	if err != nil {
 		return nil, err
@@ -66,7 +66,7 @@ func (in *Instance) BabeConfiguration() (*types.BabeConfiguration, error) {
 }
 
 // GrandpaAuthorities returns the genesis authorities from the runtime
-func (in *Instance) GrandpaAuthorities() ([]*types.Authority, error) {
+func (in *LegacyInstance) GrandpaAuthorities() ([]*types.Authority, error) {
 	ret, err := in.exec(runtime.GrandpaAuthorities, []byte{})
 	if err != nil {
 		return nil, err
@@ -81,7 +81,7 @@ func (in *Instance) GrandpaAuthorities() ([]*types.Authority, error) {
 }
 
 // ValidateTransaction runs the extrinsic through runtime function TaggedTransactionQueue_validate_transaction and returns *Validity
-func (in *Instance) ValidateTransaction(e types.Extrinsic) (*transaction.Validity, error) {
+func (in *LegacyInstance) ValidateTransaction(e types.Extrinsic) (*transaction.Validity, error) {
 	ret, err := in.exec(runtime.TaggedTransactionQueueValidateTransaction, e)
 	if err != nil {
 		return nil, err
@@ -98,7 +98,7 @@ func (in *Instance) ValidateTransaction(e types.Extrinsic) (*transaction.Validit
 }
 
 // InitializeBlock calls runtime API function Core_initialize_block
-func (in *Instance) InitializeBlock(header *types.Header) error {
+func (in *LegacyInstance) InitializeBlock(header *types.Header) error {
 	encodedHeader, err := scale.Encode(header)
 	if err != nil {
 		return fmt.Errorf("cannot encode header: %s", err)
@@ -109,17 +109,17 @@ func (in *Instance) InitializeBlock(header *types.Header) error {
 }
 
 // InherentExtrinsics calls runtime API function BlockBuilder_inherent_extrinsics
-func (in *Instance) InherentExtrinsics(data []byte) ([]byte, error) {
+func (in *LegacyInstance) InherentExtrinsics(data []byte) ([]byte, error) {
 	return in.exec(runtime.BlockBuilderInherentExtrinsics, data)
 }
 
 // ApplyExtrinsic calls runtime API function BlockBuilder_apply_extrinsic
-func (in *Instance) ApplyExtrinsic(data types.Extrinsic) ([]byte, error) {
+func (in *LegacyInstance) ApplyExtrinsic(data types.Extrinsic) ([]byte, error) {
 	return in.exec(runtime.BlockBuilderApplyExtrinsic, data)
 }
 
 // FinalizeBlock calls runtime API function BlockBuilder_finalize_block
-func (in *Instance) FinalizeBlock() (*types.Header, error) {
+func (in *LegacyInstance) FinalizeBlock() (*types.Header, error) {
 	data, err := in.exec(runtime.BlockBuilderFinalizeBlock, []byte{})
 	if err != nil {
 		return nil, err
@@ -135,7 +135,7 @@ func (in *Instance) FinalizeBlock() (*types.Header, error) {
 }
 
 // ExecuteBlock calls runtime function Core_execute_block
-func (in *Instance) ExecuteBlock(block *types.Block) ([]byte, error) {
+func (in *LegacyInstance) ExecuteBlock(block *types.Block) ([]byte, error) {
 	b := block.DeepCopy()
 
 	b.Header.Digest = [][]byte{}
@@ -146,3 +146,58 @@ func (in *Instance) ExecuteBlock(block *types.Block) ([]byte, error) {
 
 	return in.exec(runtime.CoreExecuteBlock, bdEnc)
 }
+
+// ValidateTransaction runs the extrinsic through runtime function TaggedTransactionQueue_validate_transaction and returns *Validity
+func (in *Instance) ValidateTransaction(e types.Extrinsic) (*transaction.Validity, error) {
+	return in.inst.ValidateTransaction(e)
+}
+
+// Version calls runtime function Core_Version
+func (in *Instance) Version() (*runtime.VersionAPI, error) {
+	return in.inst.Version()
+}
+
+// Metadata calls runtime function Metadata_metadata
+func (in *Instance) Metadata() ([]byte, error) {
+	return in.inst.Metadata()
+}
+
+// BabeConfiguration gets the configuration data for BABE from the runtime
+func (in *Instance) BabeConfiguration() (*types.BabeConfiguration, error) {
+	return in.inst.BabeConfiguration()
+}
+
+// GrandpaAuthorities returns the genesis authorities from the runtime
+func (in *Instance) GrandpaAuthorities() ([]*types.Authority, error) {
+	return in.inst.GrandpaAuthorities()
+}
+
+// InitializeBlock calls runtime API function Core_initialize_block
+func (in *Instance) InitializeBlock(header *types.Header) error {
+	return in.inst.InitializeBlock(header)
+}
+
+// InherentExtrinsics calls runtime API function BlockBuilder_inherent_extrinsics
+func (in *Instance) InherentExtrinsics(data []byte) ([]byte, error) {
+	return in.inst.InherentExtrinsics(data)
+}
+
+// ApplyExtrinsic calls runtime API function BlockBuilder_apply_extrinsic
+func (in *Instance) ApplyExtrinsic(data types.Extrinsic) ([]byte, error) {
+	return in.inst.ApplyExtrinsic(data)
+}
+
+// FinalizeBlock calls runtime API function BlockBuilder_finalize_block
+func (in *Instance) FinalizeBlock() (*types.Header, error) {
+	return in.inst.FinalizeBlock()
+}
+
+// ExecuteBlock calls runtime function Core_execute_block
+func (in *Instance) ExecuteBlock(block *types.Block) ([]byte, error) {
+	return in.inst.ExecuteBlock(block)
+}
+
+func (in *Instance) CheckInherents()      {}
+func (in *Instance) RandomSeed()          {}
+func (in *Instance) OffchainWorker()      {}
+func (in *Instance) GenerateSessionKeys() {}
