+++ lib/runtime/wasmtime/exports_test.go
@@ -31,7 +31,7 @@ import (
 )
 
 func TestConcurrentRuntimeCalls(t *testing.T) {
-	instance := NewTestInstance(t, runtime.NODE_RUNTIME)
+	instance := NewTestInstance(t, runtime.LEGACY_NODE_RUNTIME)
 
 	// execute 2 concurrent calls to the runtime
 	go func() {
@@ -42,7 +42,7 @@ func TestConcurrentRuntimeCalls(t *testing.T) {
 	}()
 }
 
-func TestInstance_Version_NodeRuntime(t *testing.T) {
+func TestInstance_Version_LegacyNodeRuntime(t *testing.T) {
 	// https://github.com/paritytech/substrate/blob/7b1d822446982013fa5b7ad5caff35ca84f8b7d0/core/test-runtime/src/lib.rs#L73
 	expected := &runtime.Version{
 		Spec_name:         []byte("node"),
@@ -52,7 +52,7 @@ func TestInstance_Version_NodeRuntime(t *testing.T) {
 		Impl_version:      193,
 	}
 
-	instance := NewTestInstance(t, runtime.NODE_RUNTIME)
+	instance := NewTestInstance(t, runtime.LEGACY_NODE_RUNTIME)
 	version, err := instance.Version()
 	require.NoError(t, err)
 
@@ -65,7 +65,7 @@ func TestInstance_Version_NodeRuntime(t *testing.T) {
 	require.Equal(t, expected, version.RuntimeVersion)
 }
 
-func TestInstance_BabeConfiguration_NodeRuntime(t *testing.T) {
+func TestInstance_BabeConfiguration_LegacyNodeRuntime(t *testing.T) {
 	expected := &types.BabeConfiguration{
 		SlotDuration:       3000,
 		EpochLength:        200,
@@ -75,41 +75,41 @@ func TestInstance_BabeConfiguration_NodeRuntime(t *testing.T) {
 		SecondarySlots:     true,
 	}
 
-	instance := NewTestInstance(t, runtime.NODE_RUNTIME)
+	instance := NewTestInstance(t, runtime.LEGACY_NODE_RUNTIME)
 	babeCfg, err := instance.BabeConfiguration()
 	require.NoError(t, err)
 	require.Equal(t, expected, babeCfg)
 }
 
-func TestInstance_GrandpaAuthorities_NodeRuntime(t *testing.T) {
+func TestInstance_GrandpaAuthorities_LegacyNodeRuntime(t *testing.T) {
 	expected := []*types.Authority{}
 
-	instance := NewTestInstance(t, runtime.NODE_RUNTIME)
+	instance := NewTestInstance(t, runtime.LEGACY_NODE_RUNTIME)
 	res, err := instance.GrandpaAuthorities()
 	require.NoError(t, err)
 	require.Equal(t, expected, res)
 }
 
-func TestInstance_InitializeBlock_NodeRuntime(t *testing.T) {
+func TestInstance_InitializeBlock_LegacyNodeRuntime(t *testing.T) {
 	header := &types.Header{
 		ParentHash: trie.EmptyHash,
 		Number:     big.NewInt(1),
 		Digest:     [][]byte{},
 	}
 
-	instance := NewTestInstance(t, runtime.NODE_RUNTIME)
+	instance := NewTestInstance(t, runtime.LEGACY_NODE_RUNTIME)
 	err := instance.InitializeBlock(header)
 	require.NoError(t, err)
 }
 
-func TestInstance_InherentExtrinsics_NodeRuntime(t *testing.T) {
+func TestInstance_InherentExtrinsics_LegacyNodeRuntime(t *testing.T) {
 	header := &types.Header{
 		ParentHash: trie.EmptyHash,
 		Number:     big.NewInt(1),
 		Digest:     [][]byte{},
 	}
 
-	instance := NewTestInstance(t, runtime.NODE_RUNTIME)
+	instance := NewTestInstance(t, runtime.LEGACY_NODE_RUNTIME)
 	err := instance.InitializeBlock(header)
 	require.NoError(t, err)
 
@@ -145,8 +145,8 @@ func TestInstance_InherentExtrinsics_NodeRuntime(t *testing.T) {
 	}
 }
 
-func TestInstance_FinalizeBlock_NodeRuntime(t *testing.T) {
-	instance := NewTestInstance(t, runtime.NODE_RUNTIME)
+func TestInstance_FinalizeBlock_LegacyNodeRuntime(t *testing.T) {
+	instance := NewTestInstance(t, runtime.LEGACY_NODE_RUNTIME)
 
 	header := &types.Header{
 		ParentHash: trie.EmptyHash,
@@ -206,4 +206,27 @@ func TestInstance_FinalizeBlock_NodeRuntime(t *testing.T) {
 	require.NotEqual(t, common.Hash{}, res.ExtrinsicsRoot)
 	require.NotEqual(t, trie.EmptyHash, res.StateRoot)
 	require.NotEqual(t, trie.EmptyHash, res.ExtrinsicsRoot)
+}
+
+func TestInstance_Version_NodeRuntime(t *testing.T) {
+	// https://github.com/paritytech/substrate/blob/7b1d822446982013fa5b7ad5caff35ca84f8b7d0/core/test-runtime/src/lib.rs#L73
+	expected := &runtime.Version{
+		Spec_name:         []byte("node"),
+		Impl_name:         []byte("substrate-node"),
+		Authoring_version: 10,
+		Spec_version:      193,
+		Impl_version:      193,
+	}
+
+	instance := NewTestInstance(t, runtime.NODE_RUNTIME)
+	version, err := instance.Version()
+	require.NoError(t, err)
+
+	t.Logf("Spec_name: %s\n", version.RuntimeVersion.Spec_name)
+	t.Logf("Impl_name: %s\n", version.RuntimeVersion.Impl_name)
+	t.Logf("Authoring_version: %d\n", version.RuntimeVersion.Authoring_version)
+	t.Logf("Spec_version: %d\n", version.RuntimeVersion.Spec_version)
+	t.Logf("Impl_version: %d\n", version.RuntimeVersion.Impl_version)
+
+	require.Equal(t, expected, version.RuntimeVersion)
 }
